backend/requirements.txt
pandas==2.1.4
web3==6.15.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
numpy==1.24.3 